SSD1306 for Arduino is fixed. WIP: M5 stack and SSD1306 for RPI

+++ lib/arduino/MakerPlayground/MakerPlayground.h
@@ -22,7 +22,7 @@ struct Expr {
 };
 
 void (*currentNode)(void);
-int status_code = 0;
+uint8_t status_code = 0;
 unsigned long currentTime = 0;
 unsigned long latestLogTime = 0;
 unsigned long delayEndTime = 0;
